implemented components with mobile display
