+++ app/components/Sections/ServicesInfo/index.js
