+++ app/components/Sections/index.js
@@ -10,7 +10,7 @@ const Gallery = () => {
     <div className="bg-munsell pb-[100px]">
       {/* //todo LEFT SIDE */}
 
-      <section className="p-0 bg-munsell 2xl:w-1/2 lg:w-full 2xl:fixed 2xl:left-0 lg:relative">
+      <section className="h-full bg-munsell 2xl:w-1/2 lg:w-full 2xl:fixed 2xl:left-0 lg:relative">
         <div className="w-full mt-5 ml-5 px-5 pb-3 text-aquamarine">
           <h2 className="font-bold text-2xl pt-[20px] underline">
             Made it this far?
@@ -26,7 +26,7 @@ const Gallery = () => {
       </section>
 
       {/* //todo RIGHT SIDE */}
-      <section className="bg-parchment 2xl:w-1/2 lg:w-full 2xl:absolute right-0 lg:relative overflow-scroll flex flex-col px-2 pt-2 pb-[150px]">
+      <section className="border-black border-l-2 bg-parchment 2xl:w-1/2 lg:w-full 2xl:absolute right-0 lg:relative overflow-scroll flex flex-col px-2 pt-2 pb-[150px]">
         <div className="text-onyx mb-5 w-full drop-shadow-lg pl-3 pb-3">
           <h2 className="font-bold text-2xl pt-[20px]">Gallery</h2>
           <p className="my-3 text-lg">
