+++ app/contact/page.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import ContactInfo from "../components/sections/ContactInfo";
+import ContactInfo from "../components/Sections/ContactInfo";
 
 const Contact = () => {
   return (
