+++ app/globals.css
@@ -12,7 +12,6 @@ html{
  .main_grid{
     display:grid;
     grid-template-columns: 1fr 1fr;
-    margin-top:70px;
     background: #0792a5;
  }
 
@@ -33,7 +32,7 @@ html{
 
  @media only screen and (max-width: 1024px) {
   .left-sq{
-    height:100%;
+    height:100vh;
  }
  }
